Stabilize AI setup and transcript quality
